+++ control/src/Dto/Api/Response/HelloWorldResponseDto.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Dto\Api\Response;
+
+readonly class HelloWorldResponseDto
+{
+    public function __construct(
+        public string $hello,
+        public ?array $agent = null,
+    ) { }
+}
